Create mean_var_std_calc.py
This is a calculator that can calculate the mean, variance, standard dev, max, min, sum. I initially imported Numpy (because it has built in mathematical functions)  and created a function to calculate an array of numbers and reshape it into a 3x3 matrix. From lines 11-21 is where majority of the actual calculator is being put to work. Ending my program with a dictionary to print the results and voila! I can go more into detail if anyone has questions.